SWFP-1 Migrate Spring Web Flow portlet related packages from Spring MVC to PortletMVC4Spring
