+++ webflow/pom.xml
@@ -11,11 +11,6 @@
 	<name>PortletMVC4Spring Web Flow</name>
 	<description>PortletMVC4Spring Web Flow</description>
 	
-	<properties>
-		<spring.framework.version>4.3.29.RELEASE</spring.framework.version>
-    	<webflow.version>2.4.8.RELEASE</webflow.version>
-	</properties>
-	
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.webflow</groupId>
@@ -28,10 +23,10 @@
 			<version>${webflow.version}</version>
 		</dependency>
 		<dependency>
-		    <groupId>org.springframework</groupId>
-		    <artifactId>spring-webmvc-portlet</artifactId>
-		    <version>${spring.framework.version}</version>
-		</dependency>
+			<groupId>com.liferay.portletmvc4spring</groupId>
+			<artifactId>com.liferay.portletmvc4spring.framework</artifactId>
+			<version>${project.version}</version>
+		</dependency>		
 		<dependency>
 			<groupId>javax.portlet</groupId>
 			<artifactId>portlet-api</artifactId>
