+++ webflow/src/main/java/com/liferay/portletmvc4spring/webflow/mvc/builder/FlowResourceFlowViewResolver.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.webflow.mvc.builder;
+package com.liferay.portletmvc4spring.webflow.mvc.builder;
 
 import org.springframework.context.ApplicationContext;
 import org.springframework.core.io.ContextResource;
